Address PR 711 review feedback
Security (CodeRabbit):
- Whitelist signup_host against SUPPORTED_SIGNUP_HOSTS in User model
  with inclusion validation and downcase normalization
- Validate request.host in RegistrationsController#safe_signup_host
  before assignment to prevent host header poisoning of email links

Refactor (Gemini):
- Extract duplicated default_url_options into RecipientHostRouting
  concern shared by ApplicationMailer and DeviseMailer

i18n (Gemini):
- Localize user form locale label and blank option via t()
- Localize layout default title, description, and footer site name
- Localize WebSite schema.org name in ApplicationController
- Add layout.* and users.form.* keys to ko.yml and ja.yml
